+++ android/src/main/java/id/kakzaki/blue_thermal_printer/BlueThermalPrinterPlugin.java
@@ -358,7 +358,35 @@ public class BlueThermalPrinterPlugin implements FlutterPlugin, ActivityAware,Me
           String string2 = (String) arguments.get("string2");
           int size = (int) arguments.get("size");
           String charset = (String) arguments.get("charset");
-          printLeftRight(result, string1, string2, size, charset);
+          String format = (String) arguments.get("format");
+          printLeftRight(result, string1, string2, size, charset,format);
+        } else {
+          result.error("invalid_argument", "argument 'message' not found", null);
+        }
+        break;
+      case "print3Column":
+        if (arguments.containsKey("string1")) {
+          String string1 = (String) arguments.get("string1");
+          String string2 = (String) arguments.get("string2");
+          String string3 = (String) arguments.get("string3");
+          int size = (int) arguments.get("size");
+          String charset = (String) arguments.get("charset");
+          String format = (String) arguments.get("format");
+          print3Column(result, string1, string2,string3, size, charset,format);
+        } else {
+          result.error("invalid_argument", "argument 'message' not found", null);
+        }
+        break;
+      case "print4Column":
+        if (arguments.containsKey("string1")) {
+          String string1 = (String) arguments.get("string1");
+          String string2 = (String) arguments.get("string2");
+          String string3 = (String) arguments.get("string3");
+          String string4 = (String) arguments.get("string4");
+          int size = (int) arguments.get("size");
+          String charset = (String) arguments.get("charset");
+          String format = (String) arguments.get("format");
+          print4Column(result, string1, string2,string3,string4, size, charset,format);
         } else {
           result.error("invalid_argument", "argument 'message' not found", null);
         }
@@ -574,7 +602,7 @@ public class BlueThermalPrinterPlugin implements FlutterPlugin, ActivityAware,Me
     }
   }
 
-  private void printLeftRight(Result result, String msg1, String msg2, int size ,String charset) {
+  private void printLeftRight(Result result, String msg1, String msg2, int size ,String charset,String format) {
     byte[] cc = new byte[] { 0x1B, 0x21, 0x03 }; // 0- normal size text
     // byte[] cc1 = new byte[]{0x1B,0x21,0x00}; // 0- normal size text
     byte[] bb = new byte[] { 0x1B, 0x21, 0x08 }; // 1- only bold text
@@ -605,7 +633,103 @@ public class BlueThermalPrinterPlugin implements FlutterPlugin, ActivityAware,Me
       }
       THREAD.write(PrinterCommands.ESC_ALIGN_CENTER);
       String line = String.format("%-15s %15s %n", msg1, msg2);
+      if(format != null) {
+        line = String.format(format, msg1, msg2);
+      }
+      if(charset != null) {
+        THREAD.write(line.getBytes(charset));
+      } else {
+        THREAD.write(line.getBytes());
+      }
+      result.success(true);
+    } catch (Exception ex) {
+      Log.e(TAG, ex.getMessage(), ex);
+      result.error("write_error", ex.getMessage(), exceptionToString(ex));
+    }
 
+  }
+
+  private void print3Column(Result result, String msg1, String msg2, String msg3, int size ,String charset, String format) {
+    byte[] cc = new byte[] { 0x1B, 0x21, 0x03 }; // 0- normal size text
+    // byte[] cc1 = new byte[]{0x1B,0x21,0x00}; // 0- normal size text
+    byte[] bb = new byte[] { 0x1B, 0x21, 0x08 }; // 1- only bold text
+    byte[] bb2 = new byte[] { 0x1B, 0x21, 0x20 }; // 2- bold with medium text
+    byte[] bb3 = new byte[] { 0x1B, 0x21, 0x10 }; // 3- bold with large text
+    byte[] bb4 = new byte[] { 0x1B, 0x21, 0x30 }; // 4- strong text
+    if (THREAD == null) {
+      result.error("write_error", "not connected", null);
+      return;
+    }
+    try {
+      switch (size) {
+        case 0:
+          THREAD.write(cc);
+          break;
+        case 1:
+          THREAD.write(bb);
+          break;
+        case 2:
+          THREAD.write(bb2);
+          break;
+        case 3:
+          THREAD.write(bb3);
+          break;
+        case 4:
+          THREAD.write(bb4);
+          break;
+      }
+      THREAD.write(PrinterCommands.ESC_ALIGN_CENTER);
+      String line = String.format("%-10s %10s %10s %n", msg1, msg2  , msg3);
+      if(format != null) {
+        line = String.format(format, msg1, msg2, msg3);
+      }
+      if(charset != null) {
+        THREAD.write(line.getBytes(charset));
+      } else {
+        THREAD.write(line.getBytes());
+      }
+      result.success(true);
+    } catch (Exception ex) {
+      Log.e(TAG, ex.getMessage(), ex);
+      result.error("write_error", ex.getMessage(), exceptionToString(ex));
+    }
+
+  }
+
+  private void print4Column(Result result, String msg1, String msg2,String msg3,String msg4, int size, String charset, String format) {
+    byte[] cc = new byte[] { 0x1B, 0x21, 0x03 }; // 0- normal size text
+    // byte[] cc1 = new byte[]{0x1B,0x21,0x00}; // 0- normal size text
+    byte[] bb = new byte[] { 0x1B, 0x21, 0x08 }; // 1- only bold text
+    byte[] bb2 = new byte[] { 0x1B, 0x21, 0x20 }; // 2- bold with medium text
+    byte[] bb3 = new byte[] { 0x1B, 0x21, 0x10 }; // 3- bold with large text
+    byte[] bb4 = new byte[] { 0x1B, 0x21, 0x30 }; // 4- strong text
+    if (THREAD == null) {
+      result.error("write_error", "not connected", null);
+      return;
+    }
+    try {
+      switch (size) {
+        case 0:
+          THREAD.write(cc);
+          break;
+        case 1:
+          THREAD.write(bb);
+          break;
+        case 2:
+          THREAD.write(bb2);
+          break;
+        case 3:
+          THREAD.write(bb3);
+          break;
+        case 4:
+          THREAD.write(bb4);
+          break;
+      }
+      THREAD.write(PrinterCommands.ESC_ALIGN_CENTER);
+      String line = String.format("%-8s %7s %7s %7s %n", msg1, msg2,msg3,msg4);
+      if(format != null) {
+        line = String.format(format, msg1, msg2,msg3,msg4);
+      }
       if(charset != null) {
         THREAD.write(line.getBytes(charset));
       } else {