+++ example/lib/testprint.dart
@@ -27,10 +27,18 @@ class TestPrint {
 //      bluetooth.printImageBytes(bytes.buffer.asUint8List(bytes.offsetInBytes, bytes.lengthInBytes));
         bluetooth.printLeftRight("LEFT", "RIGHT",0);
         bluetooth.printLeftRight("LEFT", "RIGHT",1);
+        bluetooth.printLeftRight("LEFT", "RIGHT",1,format: "%-15s %15s %n");
         bluetooth.printNewLine();
         bluetooth.printLeftRight("LEFT", "RIGHT",2);
         bluetooth.printLeftRight("LEFT", "RIGHT",3);
         bluetooth.printLeftRight("LEFT", "RIGHT",4);
+        bluetooth.printNewLine();
+        bluetooth.print3Column("Col1", "Col2", "Col3",1);
+        bluetooth.print3Column("Col1", "Col2", "Col3",1,format: "%-10s %10s %10s %n");
+        bluetooth.printNewLine();
+        bluetooth.print4Column("Col1","Col2","Col3","Col4",1);
+        bluetooth.print4Column("Col1","Col2","Col3","Col4",1,format: "%-8s %7s %7s %7s %n" );
+        bluetooth.printNewLine();
         String testString = " čĆžŽšŠ-H-ščđ";
         bluetooth.printCustom(testString, 1, 1, charset: "windows-1250");
         bluetooth.printLeftRight("Številka:", "18000001", 1, charset: "windows-1250");