+++ lib/blue_thermal_printer.dart
@@ -103,13 +103,38 @@ class BlueThermalPrinter {
       });
 
   Future<dynamic> printLeftRight(String string1, String string2, int size,
-          {String? charset}) =>
+          {String? charset, String? format}) =>
       _channel.invokeMethod('printLeftRight', {
         'string1': string1,
         'string2': string2,
         'size': size,
-        'charset': charset
+        'charset': charset,
+        'format':format
       });
+
+  Future<dynamic> print3Column(String string1, String string2, String string3, int size,
+      {String? charset, String? format}) =>
+      _channel.invokeMethod('print3Column', {
+        'string1': string1,
+        'string2': string2,
+        'string3': string3,
+        'size': size,
+        'charset': charset,
+        'format':format
+      });
+
+  Future<dynamic> print4Column(String string1, String string2, String string3,String string4, int size,
+      {String? charset, String? format}) =>
+      _channel.invokeMethod('print4Column', {
+        'string1': string1,
+        'string2': string2,
+        'string3': string3,
+        'string4': string4,
+        'size': size,
+        'charset': charset,
+        'format':format
+      });
+
 }
 
 class BluetoothDevice {